+++ spec/dynamoid/associations_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
+require 'spec_helper'
 
 describe Dynamoid::Associations do
   let(:magazine) { Magazine.create }